# openapi.yaml
openapi: 3.0.3
info:
  title: FHNW Connect REST API
  version: 1.0.0
  description: |
    OpenAPI specification for FHNW Connect.

    This API is derived from the uploaded README and ERD/domain model.
    It covers the core business capabilities:
    - browse and filter activities
    - view activity details
    - join and leave activities
    - favorite clubs and activities
    - manage community posts
    - create club announcements
    - inspect participation numbers

    Notes:
    - Authentication is modeled with JWT bearer tokens.
    - Role/ownership checks (student vs. club manager/admin) are left to backend authorization logic.
    - "Announcements" are modeled as community posts with category `announcement` and a `club_id`.
servers:
  - url: http://localhost:8000/api
    description: Local development
  - url: https://your-render-service.onrender.com/api
    description: Example Render deployment
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current authenticated user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/dashboard:
    get:
      tags: [Dashboard]
      summary: Get personalized home dashboard data
      description: Returns calendar activities, upcoming activities, favorite clubs, and favorite activities for the authenticated user.
      operationId: getMyDashboard
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: date_from
          schema:
            type: string
            format: date
        - in: query
          name: date_to
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Dashboard payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    User:
      type: object
      required: [user_id, username, email]
      properties:
        user_id:
          type: integer
          example: 12
        username:
          type: string
          maxLength: 50
          example: shannon
        email:
          type: string
          format: email
          maxLength: 100
          example: shannon@example.com
...
    ClubBase:
      type: object
      required: [club_name, club_description]
      properties:
        club_name:
          type: string
          maxLength: 100
          example: FHNW Chess Club
        club_description:
          type: string
          example: Weekly chess meetups and tournaments.
        image_path:
          type: string
          maxLength: 255
          nullable: true
          example: /media/clubs/chess.png
...
    ClubSummary:
      allOf:
        - type: object
          required: [club_id]
          properties:
            club_id:
              type: integer
              example: 3
        - $ref: '#/components/schemas/ClubBase'
        - type: object
          properties:
            is_favorited:
              type: boolean
              example: true
...
    ActivityCategory:
      type: string
      description: Extend this enum as your project categories become final.
      enum:
        - sports
        - club
        - workshop
        - social
        - academic
        - other
    ActivityBase:
      type: object
      required:
        - activity_title
        - activity_description
        - activity_date
        - start_time
        - end_time
        - activity_location
        - activity_category
        - max_slots
        - club_id
      properties:
        activity_title:
          type: string
          maxLength: 100
          example: Volleyball Training
        activity_description:
          type: string
          example: Weekly indoor training session for all skill levels.
        activity_date:
          type: string
          format: date
          example: '2026-05-14'
        start_time:
          type: string
          format: time
          example: '18:00:00'
        end_time:
          type: string
          format: time
          example: '20:00:00'
        activity_location:
          type: string
          maxLength: 100
          example: FHNW Gym Hall A
        max_slots:
          type: integer
          minimum: 1
          example: 24
        activity_category:
          $ref: '#/components/schemas/ActivityCategory'
        image_path:
          type: string
          maxLength: 255
          nullable: true
          example: /media/activities/volleyball.png
        club_id:
          type: integer
          example: 5
...
    ActivitySummary:
      allOf:
        - type: object
          required: [activity_id]
          properties:
            activity_id:
              type: integer
              example: 11
        - $ref: '#/components/schemas/ActivityBase'
        - type: object
          properties:
            club:
              $ref: '#/components/schemas/ClubSummary'
            joined_count:
              type: integer
              example: 18
            available_slots:
              type: integer
              example: 6
            is_favorited:
              type: boolean
              example: false
            my_registration_status:
              $ref: '#/components/schemas/RegistrationStatus'
...
    RegistrationStatus:
      type: string
      enum: [joined, left, cancelled]
      example: joined
...
    DashboardResponse:
      type: object
      required: [calendar_items, upcoming_activities, favorite_clubs, favorite_activities]
      properties:
        calendar_items:
          type: array
          items:
            $ref: '#/components/schemas/ActivitySummary'
        upcoming_activities:
          type: array
          items:
            $ref: '#/components/schemas/ActivitySummary'
        favorite_clubs:
          type: array
          items:
            $ref: '#/components/schemas/ClubSummary'
        favorite_activities:
          type: array
          items:
            $ref: '#/components/schemas/ActivitySummary'
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Resource not found
        code:
          type: string
          example: not_found
        details:
          type: object
          additionalProperties: true
          nullable: true